// ts-frontend/src/types/types.ts
export interface CustomerTypeData {
  count: number;
  acv: number;
  closed_fiscal_quarter: string;
  Cust_Type: string;
}

export interface AccountIndustryData {
  count: number;
  acv: number;
  closed_fiscal_quarter: string;
  Acct_Industry: string;
}

export interface TeamData {
  count: number;
  acv: number;
  closed_fiscal_quarter: string;
  Team: string;
}